Issue #56: Split raw xml loading into separate service.
